mysql_to_postgres_project/requirements.txt
pymysql
mysql-connector-python
sqlalchemy
pandas
sshtunnel
